Config window opens at 1/4 of primary display screen
