+++ Gobchat.App/Gobchat.csproj
@@ -548,7 +548,11 @@
     <Compile Include="Module\MemoryReader\IMemoryReaderManager.cs" />
     <Compile Include="Module\MemoryReader\Internal\FFXIVMemoryManager.cs" />
     <Compile Include="Module\Misc\Chatlogger\Internal\CustomChatLogger.cs" />
-    <Compile Include="Module\UI\BrowserAPI\Internal\BrowserAPIManager.cs" />
+    <Compile Include="Module\UI\Internal\IBrowserActorHandler.cs" />
+    <Compile Include="Module\UI\Internal\IBrowserChatHandler.cs" />
+    <Compile Include="Module\UI\Internal\IBrowserConfigHandler.cs" />
+    <Compile Include="Module\UI\Internal\IBrowserMemoryHandler.cs" />
+    <Compile Include="Module\UI\Internal\BrowserAPIManager.cs" />
     <Compile Include="Module\UI\Connector\AppModuleActorToUI.cs" />
     <Compile Include="Module\UI\Connector\AppModuleChatToUI.cs" />
     <Compile Include="Module\Chat\Internal\ChatManager.cs" />
@@ -561,20 +565,21 @@
     <Compile Include="Module\Misc\AppModuleHideOnMinimize.cs" />
     <Compile Include="Module\UI\Connector\AppModuleConfigToUI.cs" />
     <Compile Include="Module\UI\Connector\AppModuleMemoryToUI.cs" />
-    <Compile Include="Module\UI\Initialize\AppModuleLoadUI.cs" />
+    <Compile Include="Module\UI\AppModuleLoadUI.cs" />
     <Compile Include="Module\Misc\AppModuleShowHideHotkey.cs" />
-    <Compile Include="Module\UI\SynchronizeConfigWebEvent.cs" />
+    <Compile Include="Module\UI\WebEvents\SynchronizeConfigWebEvent.cs" />
     <Compile Include="Core\Chat\SystemMessageType.cs" />
-    <Compile Include="Module\UI\ToggleSearchWebEvent.cs" />
+    <Compile Include="Module\UI\WebEvents\ToggleSearchWebEvent.cs" />
     <Compile Include="Core\Util\DelayedCallback.cs" />
     <Compile Include="Module\MemoryReader\AppModuleMemoryReader.cs" />
     <Compile Include="Module\Misc\AppModuleShowConnectionOnTrayIcon.cs" />
-    <Compile Include="Module\UI\BrowserAPI\AppModuleBrowserAPIManager.cs" />
+    <Compile Include="Module\UI\AppModuleBrowserAPIManager.cs" />
     <Compile Include="Core\Util\EventHandlerManager.cs" />
     <Compile Include="Core\Util\EventHandlerConfigManager.cs" />
-    <Compile Include="Module\UI\BrowserAPI\GobchatBrowserAPI.cs" />
+    <Compile Include="Module\UI\BrowserAPI.cs" />
     <Compile Include="Module\UI\IBrowserAPIManager.cs" />
-    <Compile Include="Module\UI\UIReadChangedEventArgs.cs" />
+    <Compile Include="Module\UI\Event\UIReadChangedEventArgs.cs" />
+    <Compile Include="Module\UI\Internal\UIResourceHandler.cs" />
     <Compile Include="Module\Updater\Internal\GitHubFileDownloader.cs" />
     <Compile Include="Core\Util\IdGenerator.cs" />
     <Compile Include="Module\Cef\AppModuleCefManager.cs" />
