+++ Gobchat.App/Module/UI/AppModuleBrowserAPIManager.cs
