+++ Gobchat.App/Module/UI/Event/UIReadChangedEventArgs.cs
