+++ Gobchat.App/Module/UI/WebEvents/SynchronizeConfigWebEvent.cs
