+++ Gobchat.App/resources/ui/gobchat.js
@@ -70,15 +70,20 @@ jQuery(function ($) {
             break;
         }
     }
-    function openGobchatConfig() {
+    async function openGobchatConfig() {
         const isConfigOpen = window.localStorage.getItem(localStorageKey) || "false";
         if (isConfigOpen === "true")
             return;
         window.localStorage.setItem(localStorageKey, "true");
         gobConfig.saveToLocalStore();
-        const configWidth = gobConfig.get("behaviour.frame.config.size.width");
-        const configHeight = gobConfig.get("behaviour.frame.config.size.height");
-        const handle = window.open("config/config.html", 'Settings', `width=${configWidth},height=${configHeight}`);
+        // const configWidth = gobConfig.get("behaviour.frame.config.size.width")
+        // const configHeight = gobConfig.get("behaviour.frame.config.size.height")
+        const bounds = await GobchatAPI.getScreenDimensions();
+        const screenWidth = bounds.Item1;
+        const screenHeight = bounds.Item2;
+        const dialogWidth = screenWidth / 4;
+        const dialogHeight = screenHeight / 4;
+        const handle = window.open("config/config.html", 'Settings', `width=${dialogWidth},height=${dialogHeight}`);
         if (handle === null) {
             console.error("unable to open popup window");
             return;
